Move checkMapDataUpdate to dataActions
Move checkMapDataUpdate to dataActions along with helper functions to dataActionHelpers due to stronger relevance to dataReducer. Also rename helper functions for getWorldDataSet to differentiate helper functions for getNewRegionDataSet, and make appropriate changes to MapViewDropdown.
